FIX(gif): Adjust random selection functionality
Previously I pulled 5 gifs into an array and then picked a random number
from 0 to 5. The random number was used to select which gif out of the
list to use. Problem is I neglected to consider that gifs are 0 indexed
and that I may not always get 5 results.

I moved the number selection to after I get my GIPHY results to ensure
that I pick a random gif from my available selection.

resolves #6